+++ gateway-server/src/main/java/site/udtk/gatewayserver/config/SecurityConfig.java
@@ -1,17 +1,16 @@
-package site.udtk.globalsecurity;
+package site.udtk.gatewayserver.config;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.web.server.ServerHttpSecurity;
+import org.springframework.security.core.userdetails.MapReactiveUserDetailsService;
+import org.springframework.security.core.userdetails.ReactiveUserDetailsService;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
-import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.server.SecurityWebFilterChain;
 
 import lombok.RequiredArgsConstructor;
 
@@ -28,21 +27,21 @@ public class SecurityConfig {
 	}
 
 	@Bean
-	public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+	public SecurityWebFilterChain securityWebFilterChain(ServerHttpSecurity http) {
 		http
-			.csrf(AbstractHttpConfigurer::disable)
-			.authorizeHttpRequests(auth -> auth.anyRequest().authenticated())
+			.csrf(ServerHttpSecurity.CsrfSpec::disable)
+			.authorizeExchange(exchanges -> exchanges.anyExchange().authenticated())
 			.httpBasic(Customizer.withDefaults());
 		return http.build();
 	}
 
 	@Bean
-	public UserDetailsService userDetailsService() {
+	public ReactiveUserDetailsService userDetailsService() {
 		UserDetails user = User.builder()
 			.username(userConfig.getUsername())
 			.password(bcryptPasswordEncoder().encode(userConfig.getPassword()))
 			.roles(userConfig.getRole())
 			.build();
-		return new InMemoryUserDetailsManager(user);
+		return new MapReactiveUserDetailsService(user);
 	}
 }
